requirements.txt
RPi.GPIO
python-socketio
psycopg2
schedule
